更改cni的位置
